requirements.txt
numpy>=1.18.1
Pillow>=7.0.0
tensorflow>=2.0.0
tensorflow_hub>=0.7.0
tqdm>=4.43.0